Add additional sync mode test coverage for sub account
