+++ test/test-cases/index.js
@@ -3,7 +3,11 @@
 const apiSyncModeSqliteTestCases = require(
   './api-sync-mode-sqlite-test-cases'
 )
+const additionalApiSyncModeSqliteTestCases = require(
+  './additional-api-sync-mode-sqlite-test-cases'
+)
 
 module.exports = {
-  apiSyncModeSqliteTestCases
+  apiSyncModeSqliteTestCases,
+  additionalApiSyncModeSqliteTestCases
 }
